+++ CHANGELOG.md
@@ -59,6 +59,9 @@ If switching from Carbon,
     `Grant-CServicePermission -Name $svcName -PrincipalName $user -QueryConfig -EnumerateDependents` would be changed
     to `Grant-CServicePermission -Name $svcName -PrincipalName $user 'QueryConfig, EnumerateDependents'`.
 * Rename usages of the `Revoke-CServicePermission` function's `Identity` parameter to `PrincipalName`.
+* `Grant-CServiceControlPermission`:
+  * Rename usages of the `ServiceName` parameter to `Name`.
+  * Rename usages of the `Identity` parameter to `PrincipalName`.
 
 ### Added
 
@@ -88,7 +91,9 @@ If switching from Carbon,
 * `Revoke-CServicePermission`:
   * `WhatIf` support.
   * writes an information message when revoking permission.
-
+* `Grant-CServiceControlPermission`:
+  * `WhatIf support
+  * writes an information message when granting permission.
 
 ### Changed
 
@@ -116,6 +121,9 @@ The following changed from Carbon:
     `Permission` as `'QueryConfig, EnumerateDependents'`.
   * only sets permissions if they are different than existing permissions
 * Renamed the `Revoke-CServicePermission` function's `Identity` parameter to `PrincipalName`.
+* `Grant-CServiceControlPermission`:
+  * Renamed the `ServiceName` parameter to `Name`.
+  * Renamed the `Identity` parameter to `PrincipalName`.
 
 ### Fixed